Initial commit for IN-DEV
**IN-DEV Builds / Experimental builds are live!** 
Be warned that things will break & it's highly experimental.

If you see any IDs in the code, please create an issue or a pull request and/or remove it. 
We'll start storing owner IDs on a database soon. Please be careful on this branch.

As always we're an open-source collective and we'd love for you to contribute to the code.

If you need any help or want to debug with us, consider talking to us in #campus-grounds on Discord!